Updated import paths, better make, specify branch to checkout
THe import paths have to be updated since the fork is in a different
repository path.

The makefile's install command now runs go install. The build command
outputs a binary named goreportcard. The old install command is
now install-deps to better reflect its purpose.

The branch to checkout out can be given on the command line using
the "-branch" flag. By default it's whatever is the default branch
of the repo, typically master. Eventually, the branch will be specified
in the web gui.